Updated Enqueueing and Category_Slugs
The theme has been updated to properly enqueue parent and child css and some JS so the child theme can use the original version of the Colleges Theme without having to mod it.

Addition of registering category_slugs so the feed from the site can be used in the UCF News plugin.