Addition of Search
Added a custom search form with Athena and Bootstrap styling that can be envoked using a shortcode.

Excluded the UCF Section CPT from search results as they will be included in the page results.

Added a function that adds Athena and Bootstrap styling to pagination.

Created a styled, custom search.php page.