Simplified Excerpt for the Modern Post Layout
I simplified getting the post excerpt for the Modern layout. In this layout I have the excerpt length hard coded so a custom function is not needed. Also, the previous code would throw a PHP warning if the field wasn't defined.

I'm now using the built in WordPress function get_the_excerpt() instead.